Change function of form submit button to submit
The form didn't have ability to validate the fields as the submit button
was bypassing the form validation step, even when the fields are set to
required. Change to function of the button to submit and then
over-riding the default behavior of submit when it's triggered allows
for form validation while adding the data to the table, as opposed to an
external database.